requirements.txt
Django==3.2
gunicorn==20.1.0
python-dotenv==0.20.0
mssql-django==1.1.3
pyodbc==4.0.34
whitenoise==6.2.0
djangorestframework==3.13.1
# pyjwt==2.4.0
django-crispy-forms==1.14.0
django-cors-headers==3.13.0
django-storages==1.13.1
azure-storage-blob==12.13.1
azure-storage-file-share==12.9.0
PyPDF2==2.11.1
python-docx==0.8.11
cloudconvert==2.1.0
django-cors-headers==3.13.0
# textract==1.6.5
# pdfreader==0.1.12
# antiword-0.1.0
# django-3-jet==1.0.8
# django-admin-interface==0.22.0




